atzmo/requirements.txt
Pillow

playwright
